requirements.txt
selenium
imageio
imageio==2.4.1
moviepy
xmltodict
websocket
websocket-client
# baidu ai
baidu-aip
pygame
playsound
pyobjc
# ocr识别
pillow
pytesseract